Read the base address of the LFB from BAR0 of the PCI device.

+++ io.h
@@ -3,26 +3,26 @@
 
 #include "stdint.h"
 
-static inline uint16_t inw(unsigned port)
+static inline uint16_t inw(uint16_t port)
 {
     uint16_t data;
     asm volatile ("inw %w1, %w0" : "=a" (data) : "Nd" (port));
     return data;
 }
 
-static inline void outw(uint16_t data, unsigned port)
+static inline void outw(uint16_t data, uint16_t port)
 {
     asm volatile ("outw %w0, %w1" : : "a" (data), "Nd" (port));
 }
 
-static inline uint32_t inl(unsigned port)
+static inline uint32_t inl(uint16_t port)
 {
     uint32_t data;
     asm volatile ("inl %w1, %0" : "=a" (data) : "Nd" (port));
     return data;
 }
 
-static inline void outl(uint32_t data, unsigned port)
+static inline void outl(uint32_t data, uint16_t port)
 {
     asm volatile ("outl %0, %w1" : : "a" (data), "Nd" (port));
 }
